requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.2

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
# Note: sqlite3 is built into Python

# ML and NLP (CPU-only versions)
# NOTE: Create a pip.conf or use --extra-index-url at install time for CPU torch:
#   pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cpu
sentence-transformers>=2.3.0
torch>=2.2.0,<2.5.0
transformers==4.35.2
numpy>=1.24.4,<2.0.0
scikit-learn>=1.3.2

# PDF processing
pypdf==3.17.1
python-multipart==0.0.6

# Text processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Utilities
python-dotenv==1.0.0
requests>=2.31.0
tqdm>=4.66.1
pandas>=2.1.4

# Logging and monitoring
loguru==0.7.2